Added sanity check for IP
At boot-up the program might not be able to get the IP-address
(unless connected already). We thus need to make sure that is a
sane IP. Also IP might change while program is running.

Current implementation checks every 10 seconds, this might be
overkill. It is possible to check more seldom unless the current
ip-info is unknown (ie after boot-up)